+++ src/Oro/Bundle/AddressBundle/Provider/PhoneProvider.php
@@ -307,7 +307,7 @@ class PhoneProvider implements PhoneProviderInterface, ResetInterface
         if (null === $this->sortedTargetEntities) {
             ksort($this->targetEntities);
             $this->sortedTargetEntities = $this->targetEntities
-                ? array_merge(...$this->targetEntities)
+                ? array_merge(...array_values($this->targetEntities))
                 : [];
         }