+++ src/Oro/Bundle/AssetBundle/Command/OroAssetsBuildCommand.php
@@ -96,6 +96,7 @@ class OroAssetsBuildCommand extends Command
             ->addOption('skip-js', null, InputOption::VALUE_NONE, 'Skip build of JS assets')
             ->addOption('with-babel', null, InputOption::VALUE_NONE, 'Transpile code with Babel')
             ->addOption('skip-sourcemap', null, InputOption::VALUE_NONE, 'Skip building source maps')
+            ->addOption('skip-rtl', null, InputOption::VALUE_NONE, 'Skip building RTL styles')
             ->addOption('analyze', null, InputOption::VALUE_NONE, 'Run BundleAnalyzerPlugin')
         ;
         $this
@@ -150,13 +151,14 @@ in <comment>vendor/oro/platform/build</comment> folder. It may be required when
 
   <info>php %command.full_name% --npm-install</info>
 
-The <info>--skip-css</info>, <info>--skip-js</info>, <info>--with-babel</info> and <info>--skip-sourcemap</info> options allow to
-skip building CSS and JavaScript files, and transpiling Javascript with Babel
-and skip building sourcemaps respectively:
+The <info>--skip-css</info>, <info>--skip-js</info>, <info>--with-babel</info>, <info>--skip-sourcemap</info> and <info>--skip-rtl</info> options allow to
+skip building CSS and JavaScript files, and transpiling Javascript with Babel,
+skip building sourcemaps and skip building RTL styles respectively:
 
   <info>php %command.full_name% --skip-css</info>
   <info>php %command.full_name% --skip-js</info>
   <info>php %command.full_name% --skip-sourcemap</info>
+  <info>php %command.full_name% --skip-rtl</info>
   <info>php %command.full_name% --with-babel</info>
 
 The <info>--analyze</info> option can be used to run BundleAnalyzerPlugin:
@@ -185,6 +187,7 @@ HELP
             ->addUsage('--skip-css')
             ->addUsage('--skip-js')
             ->addUsage('--skip-sourcemap')
+            ->addUsage('--skip-rtl')
             ->addUsage('--with-babel')
             ->addUsage('--analyze')
         ;
@@ -296,6 +299,9 @@ HELP
         }
         if ($input->getOption('skip-sourcemap')) {
             $command[] = '--env.skipSourcemap';
+        }
+        if ($input->getOption('skip-rtl')) {
+            $command[] = '--env.skipRTL';
         }
         if ($input->getOption('analyze')) {
             $command[] = '--env.analyze';