+++ src/Oro/Component/Layout/LayoutRegistry.php
@@ -6,6 +6,9 @@ use Oro\Component\Layout\Block\OptionsResolver\OptionsResolver;
 use Oro\Component\Layout\Block\Type\Options;
 use Oro\Component\Layout\Extension\ExtensionInterface;
 
+/**
+ * The main registry of the Layout component
+ */
 class LayoutRegistry implements LayoutRegistryInterface
 {
     /** @var array */
@@ -218,7 +221,7 @@ class LayoutRegistry implements LayoutRegistryInterface
         if (null === $this->sorted) {
             ksort($this->extensions);
             $this->sorted = !empty($this->extensions)
-                ? call_user_func_array('array_merge', $this->extensions)
+                ? array_merge(...array_values($this->extensions))
                 : [];
         }